requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
PyYAML>=6.0

# Testing dependencies
pytest>=7.4.0
hypothesis>=6.82.0
pytest-cov>=4.1.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0